chore: License tab -> stockImg tab
