+++ src/component/order/editor/icons/index.js
@@ -3,7 +3,6 @@ import Uploads from './Uploads';
 import SwapHorizontal from './SwapHorizontal';
 import Layers from './Layers';
 import AI from './AI';
-import License from './License';
 import Text from './Text';
 class Icons {
   static Images = Images;
@@ -11,7 +10,6 @@ class Icons {
   static Customize = SwapHorizontal;
   static Layers = Layers;
   static AI = AI;
-  static License = License;
   static Text = Text;
 }
