+++ src/constants/mock-data.js
@@ -1,396 +1,3 @@
-export const images = [
-  {
-    id: 3493777,
-    width: 3000,
-    height: 3000,
-    url: 'https://www.pexels.com/photo/photo-of-bridge-under-white-clouds-3493777/',
-    photographer: 'Aleksey Kuprikov',
-    photographer_url: 'https://www.pexels.com/@aleksey-kuprikov-1883853',
-    photographer_id: 1883853,
-    avg_color: '#6B7874',
-    src: {
-      original: 'https://images.pexels.com/photos/3493777/pexels-photo-3493777.jpeg',
-      large2x:
-        'https://images.pexels.com/photos/3493777/pexels-photo-3493777.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=650&w=940',
-      large:
-        'https://images.pexels.com/photos/3493777/pexels-photo-3493777.jpeg?auto=compress&cs=tinysrgb&h=650&w=940',
-      medium:
-        'https://images.pexels.com/photos/3493777/pexels-photo-3493777.jpeg?auto=compress&cs=tinysrgb&h=350',
-      small:
-        'https://images.pexels.com/photos/3493777/pexels-photo-3493777.jpeg?auto=compress&cs=tinysrgb&h=130',
-      portrait:
-        'https://images.pexels.com/photos/3493777/pexels-photo-3493777.jpeg?auto=compress&cs=tinysrgb&fit=crop&h=1200&w=800',
-      landscape:
-        'https://images.pexels.com/photos/3493777/pexels-photo-3493777.jpeg?auto=compress&cs=tinysrgb&fit=crop&h=627&w=1200',
-      tiny: 'https://images.pexels.com/photos/3493777/pexels-photo-3493777.jpeg?auto=compress&cs=tinysrgb&dpr=1&fit=crop&h=200&w=280',
-    },
-    liked: false,
-    alt: 'Photo of Bridge Under White Clouds',
-  },
-  {
-    id: 1683492,
-    width: 3648,
-    height: 3648,
-    url: 'https://www.pexels.com/photo/ocean-wave-splashing-on-dock-1683492/',
-    photographer: 'Ricardo Esquivel',
-    photographer_url: 'https://www.pexels.com/@rickyrecap',
-    photographer_id: 722822,
-    avg_color: '#8298A9',
-    src: {
-      original: 'https://images.pexels.com/photos/1683492/pexels-photo-1683492.jpeg',
-      large2x:
-        'https://images.pexels.com/photos/1683492/pexels-photo-1683492.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=650&w=940',
-      large:
-        'https://images.pexels.com/photos/1683492/pexels-photo-1683492.jpeg?auto=compress&cs=tinysrgb&h=650&w=940',
-      medium:
-        'https://images.pexels.com/photos/1683492/pexels-photo-1683492.jpeg?auto=compress&cs=tinysrgb&h=350',
-      small:
-        'https://images.pexels.com/photos/1683492/pexels-photo-1683492.jpeg?auto=compress&cs=tinysrgb&h=130',
-      portrait:
-        'https://images.pexels.com/photos/1683492/pexels-photo-1683492.jpeg?auto=compress&cs=tinysrgb&fit=crop&h=1200&w=800',
-      landscape:
-        'https://images.pexels.com/photos/1683492/pexels-photo-1683492.jpeg?auto=compress&cs=tinysrgb&fit=crop&h=627&w=1200',
-      tiny: 'https://images.pexels.com/photos/1683492/pexels-photo-1683492.jpeg?auto=compress&cs=tinysrgb&dpr=1&fit=crop&h=200&w=280',
-    },
-    liked: false,
-    alt: 'Ocean Wave Splashing on Dock',
-  },
-  {
-    id: 789555,
-    width: 3327,
-    height: 3327,
-    url: 'https://www.pexels.com/photo/woman-in-black-hijab-headscarf-walking-on-field-789555/',
-    photographer: 'Samuel Theo Manat Silitonga',
-    photographer_url: 'https://www.pexels.com/@samsilitongajr',
-    photographer_id: 218404,
-    avg_color: '#A19584',
-    src: {
-      original: 'https://images.pexels.com/photos/789555/pexels-photo-789555.jpeg',
-      large2x:
-        'https://images.pexels.com/photos/789555/pexels-photo-789555.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=650&w=940',
-      large:
-        'https://images.pexels.com/photos/789555/pexels-photo-789555.jpeg?auto=compress&cs=tinysrgb&h=650&w=940',
-      medium:
-        'https://images.pexels.com/photos/789555/pexels-photo-789555.jpeg?auto=compress&cs=tinysrgb&h=350',
-      small:
-        'https://images.pexels.com/photos/789555/pexels-photo-789555.jpeg?auto=compress&cs=tinysrgb&h=130',
-      portrait:
-        'https://images.pexels.com/photos/789555/pexels-photo-789555.jpeg?auto=compress&cs=tinysrgb&fit=crop&h=1200&w=800',
-      landscape:
-        'https://images.pexels.com/photos/789555/pexels-photo-789555.jpeg?auto=compress&cs=tinysrgb&fit=crop&h=627&w=1200',
-      tiny: 'https://images.pexels.com/photos/789555/pexels-photo-789555.jpeg?auto=compress&cs=tinysrgb&dpr=1&fit=crop&h=200&w=280',
-    },
-    liked: false,
-    alt: 'Woman in Black Hijab Headscarf Walking on Field',
-  },
-  {
-    id: 2894944,
-    width: 2359,
-    height: 2359,
-    url: 'https://www.pexels.com/photo/man-standing-on-grass-field-2894944/',
-    photographer: 'Gagaz Adam',
-    photographer_url: 'https://www.pexels.com/@ggzadam',
-    photographer_id: 398931,
-    avg_color: '#4A5B61',
-    src: {
-      original: 'https://images.pexels.com/photos/2894944/pexels-photo-2894944.jpeg',
-      large2x:
-        'https://images.pexels.com/photos/2894944/pexels-photo-2894944.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=650&w=940',
-      large:
-        'https://images.pexels.com/photos/2894944/pexels-photo-2894944.jpeg?auto=compress&cs=tinysrgb&h=650&w=940',
-      medium:
-        'https://images.pexels.com/photos/2894944/pexels-photo-2894944.jpeg?auto=compress&cs=tinysrgb&h=350',
-      small:
-        'https://images.pexels.com/photos/2894944/pexels-photo-2894944.jpeg?auto=compress&cs=tinysrgb&h=130',
-      portrait:
-        'https://images.pexels.com/photos/2894944/pexels-photo-2894944.jpeg?auto=compress&cs=tinysrgb&fit=crop&h=627&w=1200',
-      tiny: 'https://images.pexels.com/photos/2894944/pexels-photo-2894944.jpeg?auto=compress&cs=tinysrgb&dpr=1&fit=crop&h=200&w=280',
-    },
-    liked: false,
-    alt: 'Man Standing on Grass Field',
-  },
-  {
-    id: 2437286,
-    width: 4681,
-    height: 4681,
-    url: 'https://www.pexels.com/photo/photo-of-snow-capped-mountain-during-evening-2437286/',
-    photographer: 'eberhard grossgasteiger',
-    photographer_url: 'https://www.pexels.com/@eberhardgross',
-    photographer_id: 121938,
-    avg_color: '#3B3B41',
-    src: {
-      original: 'https://images.pexels.com/photos/2437286/pexels-photo-2437286.jpeg',
-      large2x:
-        'https://images.pexels.com/photos/2437286/pexels-photo-2437286.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=650&w=940',
-      large:
-        'https://images.pexels.com/photos/2437286/pexels-photo-2437286.jpeg?auto=compress&cs=tinysrgb&h=650&w=940',
-      medium:
-        'https://images.pexels.com/photos/2437286/pexels-photo-2437286.jpeg?auto=compress&cs=tinysrgb&h=350',
-      small:
-        'https://images.pexels.com/photos/2437286/pexels-photo-2437286.jpeg?auto=compress&cs=tinysrgb&h=130',
-      portrait:
-        'https://images.pexels.com/photos/2437286/pexels-photo-2437286.jpeg?auto=compress&cs=tinysrgb&fit=crop&h=1200&w=800',
-      landscape:
-        'https://images.pexels.com/photos/2437286/pexels-photo-2437286.jpeg?auto=compress&cs=tinysrgb&fit=crop&h=627&w=1200',
-      tiny: 'https://images.pexels.com/photos/2437286/pexels-photo-2437286.jpeg?auto=compress&cs=tinysrgb&dpr=1&fit=crop&h=200&w=280',
-    },
-    liked: false,
-    alt: 'Photo of Snow Capped Mountain During Evening',
-  },
-  {
-    id: 670741,
-    width: 6534,
-    height: 6534,
-    url: 'https://www.pexels.com/photo/purple-and-yellow-petal-flower-670741/',
-    photographer: 'Cindy Gustafson',
-    photographer_url: 'https://www.pexels.com/@cindyg',
-    photographer_id: 206336,
-    avg_color: '#816563',
-    src: {
-      original: 'https://images.pexels.com/photos/670741/pexels-photo-670741.jpeg',
-      large2x:
-        'https://images.pexels.com/photos/670741/pexels-photo-670741.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=650&w=940',
-      large:
-        'https://images.pexels.com/photos/670741/pexels-photo-670741.jpeg?auto=compress&cs=tinysrgb&h=650&w=940',
-      medium:
-        'https://images.pexels.com/photos/670741/pexels-photo-670741.jpeg?auto=compress&cs=tinysrgb&h=350',
-      small:
-        'https://images.pexels.com/photos/670741/pexels-photo-670741.jpeg?auto=compress&cs=tinysrgb&h=130',
-      portrait:
-        'https://images.pexels.com/photos/670741/pexels-photo-670741.jpeg?auto=compress&cs=tinysrgb&fit=crop&h=1200&w=800',
-      landscape:
-        'https://images.pexels.com/photos/670741/pexels-photo-670741.jpeg?auto=compress&cs=tinysrgb&fit=crop&h=627&w=1200',
-      tiny: 'https://images.pexels.com/photos/670741/pexels-photo-670741.jpeg?auto=compress&cs=tinysrgb&dpr=1&fit=crop&h=200&w=280',
-    },
-    liked: false,
-    alt: 'Purple and Yellow Petal Flower',
-  },
-  {
-    id: 105808,
-    width: 2287,
-    height: 2287,
-    url: 'https://www.pexels.com/photo/blue-orange-and-green-bird-on-yellow-flower-105808/',
-    photographer: 'mark broadhurst',
-    photographer_url: 'https://www.pexels.com/@markb',
-    photographer_id: 25197,
-    avg_color: '#72685B',
-    src: {
-      original: 'https://images.pexels.com/photos/105808/pexels-photo-105808.jpeg',
-      large2x:
-        'https://images.pexels.com/photos/105808/pexels-photo-105808.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=650&w=940',
-      large:
-        'https://images.pexels.com/photos/105808/pexels-photo-105808.jpeg?auto=compress&cs=tinysrgb&h=650&w=940',
-      medium:
-        'https://images.pexels.com/photos/105808/pexels-photo-105808.jpeg?auto=compress&cs=tinysrgb&h=350',
-      small:
-        'https://images.pexels.com/photos/105808/pexels-photo-105808.jpeg?auto=compress&cs=tinysrgb&h=130',
-      portrait:
-        'https://images.pexels.com/photos/105808/pexels-photo-105808.jpeg?auto=compress&cs=tinysrgb&fit=crop&h=1200&w=800',
-      landscape:
-        'https://images.pexels.com/photos/105808/pexels-photo-105808.jpeg?auto=compress&cs=tinysrgb&fit=crop&h=627&w=1200',
-      tiny: 'https://images.pexels.com/photos/105808/pexels-photo-105808.jpeg?auto=compress&cs=tinysrgb&dpr=1&fit=crop&h=200&w=280',
-    },
-    liked: false,
-    alt: 'Blue Orange and Green Bird on Yellow Flower',
-  },
-  {
-    id: 694587,
-    width: 3780,
-    height: 3780,
-    url: 'https://www.pexels.com/photo/silhouette-of-boy-running-in-body-of-water-during-sunset-694587/',
-    photographer: 'Samuel Theo Manat Silitonga',
-    photographer_url: 'https://www.pexels.com/@samsilitongajr',
-    photographer_id: 218404,
-    avg_color: '#CBC6B8',
-    src: {
-      original: 'https://images.pexels.com/photos/694587/pexels-photo-694587.jpeg',
-      large2x:
-        'https://images.pexels.com/photos/694587/pexels-photo-694587.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=650&w=940',
-      large:
-        'https://images.pexels.com/photos/694587/pexels-photo-694587.jpeg?auto=compress&cs=tinysrgb&h=650&w=940',
-      medium:
-        'https://images.pexels.com/photos/694587/pexels-photo-694587.jpeg?auto=compress&cs=tinysrgb&h=350',
-      small:
-        'https://images.pexels.com/photos/694587/pexels-photo-694587.jpeg?auto=compress&cs=tinysrgb&h=130',
-      portrait:
-        'https://images.pexels.com/photos/694587/pexels-photo-694587.jpeg?auto=compress&cs=tinysrgb&fit=crop&h=1200&w=800',
-      landscape:
-        'https://images.pexels.com/photos/694587/pexels-photo-694587.jpeg?auto=compress&cs=tinysrgb&fit=crop&h=627&w=1200',
-      tiny: 'https://images.pexels.com/photos/694587/pexels-photo-694587.jpeg?auto=compress&cs=tinysrgb&dpr=1&fit=crop&h=200&w=280',
-    },
-    liked: false,
-    alt: 'Silhouette of Boy Running in Body of Water during Sunset',
-  },
-  {
-    id: 4275360,
-    width: 2048,
-    height: 2048,
-    url: 'https://www.pexels.com/photo/colorful-butterfly-on-stem-of-dry-plant-4275360/',
-    photographer: 'Pavel Hájek',
-    photographer_url: 'https://www.pexels.com/@pavel-hajek-2426651',
-    photographer_id: 2426651,
-    avg_color: '#8B8E3D',
-    src: {
-      original: 'https://images.pexels.com/photos/4275360/pexels-photo-4275360.jpeg',
-      large2x:
-        'https://images.pexels.com/photos/4275360/pexels-photo-4275360.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=650&w=940',
-      large:
-        'https://images.pexels.com/photos/4275360/pexels-photo-4275360.jpeg?auto=compress&cs=tinysrgb&h=650&w=940',
-      medium:
-        'https://images.pexels.com/photos/4275360/pexels-photo-4275360.jpeg?auto=compress&cs=tinysrgb&h=350',
-      small:
-        'https://images.pexels.com/photos/4275360/pexels-photo-4275360.jpeg?auto=compress&cs=tinysrgb&h=130',
-      portrait:
-        'https://images.pexels.com/photos/4275360/pexels-photo-4275360.jpeg?auto=compress&cs=tinysrgb&fit=crop&h=1200&w=800',
-      landscape:
-        'https://images.pexels.com/photos/4275360/pexels-photo-4275360.jpeg?auto=compress&cs=tinysrgb&fit=crop&h=627&w=1200',
-      tiny: 'https://images.pexels.com/photos/4275360/pexels-photo-4275360.jpeg?auto=compress&cs=tinysrgb&dpr=1&fit=crop&h=200&w=280',
-    },
-    liked: false,
-    alt: 'Small common blue butterfly sitting on thin twig in wildlife against blurred green background',
-  },
-  {
-    id: 7494171,
-    width: 2054,
-    height: 2054,
-    url: 'https://www.pexels.com/photo/close-up-shot-of-a-butterfly-on-a-pink-flower-7494171/',
-    photographer: 'Leon Huang',
-    photographer_url: 'https://www.pexels.com/@leon-huang-37073985',
-    photographer_id: 37073985,
-    avg_color: '#85B057',
-    src: {
-      original: 'https://images.pexels.com/photos/7494171/pexels-photo-7494171.jpeg',
-      large2x:
-        'https://images.pexels.com/photos/7494171/pexels-photo-7494171.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=650&w=940',
-      large:
-        'https://images.pexels.com/photos/7494171/pexels-photo-7494171.jpeg?auto=compress&cs=tinysrgb&h=650&w=940',
-      medium:
-        'https://images.pexels.com/photos/7494171/pexels-photo-7494171.jpeg?auto=compress&cs=tinysrgb&h=350',
-      small:
-        'https://images.pexels.com/photos/7494171/pexels-photo-7494171.jpeg?auto=compress&cs=tinysrgb&h=130',
-      portrait:
-        'https://images.pexels.com/photos/7494171/pexels-photo-7494171.jpeg?auto=compress&cs=tinysrgb&fit=crop&h=1200&w=800',
-      landscape:
-        'https://images.pexels.com/photos/7494171/pexels-photo-7494171.jpeg?auto=compress&cs=tinysrgb&fit=crop&h=627&w=1200',
-      tiny: 'https://images.pexels.com/photos/7494171/pexels-photo-7494171.jpeg?auto=compress&cs=tinysrgb&dpr=1&fit=crop&h=200&w=280',
-    },
-    liked: false,
-    alt: 'Close-Up Shot of a Butterfly on a Pink Flower',
-  },
-  {
-    id: 5846615,
-    width: 3024,
-    height: 3024,
-    url: 'https://www.pexels.com/photo/japanese-torii-gate-in-forest-5846615/',
-    photographer: 'Ayako S',
-    photographer_url: 'https://www.pexels.com/@ayako-s',
-    photographer_id: 3771433,
-    avg_color: '#525B4E',
-    src: {
-      original: 'https://images.pexels.com/photos/5846615/pexels-photo-5846615.jpeg',
-      large2x:
-        'https://images.pexels.com/photos/5846615/pexels-photo-5846615.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=650&w=940',
-      large:
-        'https://images.pexels.com/photos/5846615/pexels-photo-5846615.jpeg?auto=compress&cs=tinysrgb&h=650&w=940',
-      medium:
-        'https://images.pexels.com/photos/5846615/pexels-photo-5846615.jpeg?auto=compress&cs=tinysrgb&h=350',
-      small:
-        'https://images.pexels.com/photos/5846615/pexels-photo-5846615.jpeg?auto=compress&cs=tinysrgb&h=130',
-      portrait:
-        'https://images.pexels.com/photos/5846615/pexels-photo-5846615.jpeg?auto=compress&cs=tinysrgb&fit=crop&h=1200&w=800',
-      landscape:
-        'https://images.pexels.com/photos/5846615/pexels-photo-5846615.jpeg?auto=compress&cs=tinysrgb&fit=crop&h=627&w=1200',
-      tiny: 'https://images.pexels.com/photos/5846615/pexels-photo-5846615.jpeg?auto=compress&cs=tinysrgb&dpr=1&fit=crop&h=200&w=280',
-    },
-    liked: false,
-    alt: 'Japanese Torii Gate in Forest',
-  },
-  {
-    id: 5840879,
-    width: 2848,
-    height: 2848,
-    url: 'https://www.pexels.com/photo/legs-of-couple-lying-on-grass-5840879/',
-    photographer: 'Scott Webb',
-    photographer_url: 'https://www.pexels.com/@scottwebb',
-    photographer_id: 39047,
-    avg_color: '#759755',
-    src: {
-      original: 'https://images.pexels.com/photos/5840879/pexels-photo-5840879.jpeg',
-      large2x:
-        'https://images.pexels.com/photos/5840879/pexels-photo-5840879.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=650&w=940',
-      large:
-        'https://images.pexels.com/photos/5840879/pexels-photo-5840879.jpeg?auto=compress&cs=tinysrgb&h=650&w=940',
-      medium:
-        'https://images.pexels.com/photos/5840879/pexels-photo-5840879.jpeg?auto=compress&cs=tinysrgb&h=350',
-      small:
-        'https://images.pexels.com/photos/5840879/pexels-photo-5840879.jpeg?auto=compress&cs=tinysrgb&h=130',
-      portrait:
-        'https://images.pexels.com/photos/5840879/pexels-photo-5840879.jpeg?auto=compress&cs=tinysrgb&fit=crop&h=1200&w=800',
-      landscape:
-        'https://images.pexels.com/photos/5840879/pexels-photo-5840879.jpeg?auto=compress&cs=tinysrgb&fit=crop&h=627&w=1200',
-      tiny: 'https://images.pexels.com/photos/5840879/pexels-photo-5840879.jpeg?auto=compress&cs=tinysrgb&dpr=1&fit=crop&h=200&w=280',
-    },
-    liked: false,
-    alt: 'Woman in Blue Denim Jeans Lying on White Textile on Green Grass Field',
-  },
-  {
-    id: 5815904,
-    width: 3000,
-    height: 3000,
-    url: 'https://www.pexels.com/photo/path-in-field-5815904/',
-    photographer: 'Razvan Constantin',
-    photographer_url: 'https://www.pexels.com/@parad0xx',
-    photographer_id: 3906129,
-    avg_color: '#8C8D7B',
-    src: {
-      original: 'https://images.pexels.com/photos/5815904/pexels-photo-5815904.jpeg',
-      large2x:
-        'https://images.pexels.com/photos/5815904/pexels-photo-5815904.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=650&w=940',
-      large:
-        'https://images.pexels.com/photos/5815904/pexels-photo-5815904.jpeg?auto=compress&cs=tinysrgb&h=650&w=940',
-      medium:
-        'https://images.pexels.com/photos/5815904/pexels-photo-5815904.jpeg?auto=compress&cs=tinysrgb&h=350',
-      small:
-        'https://images.pexels.com/photos/5815904/pexels-photo-5815904.jpeg?auto=compress&cs=tinysrgb&h=130',
-      portrait:
-        'https://images.pexels.com/photos/5815904/pexels-photo-5815904.jpeg?auto=compress&cs=tinysrgb&fit=crop&h=1200&w=800',
-      landscape:
-        'https://images.pexels.com/photos/5815904/pexels-photo-5815904.jpeg?auto=compress&cs=tinysrgb&fit=crop&h=627&w=1200',
-      tiny: 'https://images.pexels.com/photos/5815904/pexels-photo-5815904.jpeg?auto=compress&cs=tinysrgb&dpr=1&fit=crop&h=200&w=280',
-    },
-    liked: false,
-    alt: 'Path in Field',
-  },
-  {
-    id: 12736377,
-    width: 2160,
-    height: 2160,
-    url: 'https://www.pexels.com/photo/leafless-tree-in-the-middle-of-a-lake-12736377/',
-    photographer: 'Mitchell Henderson',
-    photographer_url: 'https://www.pexels.com/@mtk402',
-    photographer_id: 425498,
-    avg_color: '#AAB3C2',
-    src: {
-      original: 'https://images.pexels.com/photos/12736377/pexels-photo-12736377.jpeg',
-      large2x:
-        'https://images.pexels.com/photos/12736377/pexels-photo-12736377.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=650&w=940',
-      large:
-        'https://images.pexels.com/photos/12736377/pexels-photo-12736377.jpeg?auto=compress&cs=tinysrgb&h=650&w=940',
-      medium:
-        'https://images.pexels.com/photos/12736377/pexels-photo-12736377.jpeg?auto=compress&cs=tinysrgb&h=350',
-      small:
-        'https://images.pexels.com/photos/12736377/pexels-photo-12736377.jpeg?auto=compress&cs=tinysrgb&h=130',
-      portrait:
-        'https://images.pexels.com/photos/12736377/pexels-photo-12736377.jpeg?auto=compress&cs=tinysrgb&fit=crop&h=1200&w=800',
-      landscape:
-        'https://images.pexels.com/photos/12736377/pexels-photo-12736377.jpeg?auto=compress&cs=tinysrgb&fit=crop&h=627&w=1200',
-      tiny: 'https://images.pexels.com/photos/12736377/pexels-photo-12736377.jpeg?auto=compress&cs=tinysrgb&dpr=1&fit=crop&h=200&w=280',
-    },
-    liked: false,
-    alt: '',
-  },
-];
-
 export const sampleFrames = [
   {
     id: 1,
